Add a missing _update_preedit()
There was a bug when a minimum numbers of characters for completion
was used together with the option "Use preedit style only if lookup
enabled":

When the preedit is longer than the limit, lookup is enabled and the
preedit should be styled (underlined). If after typing backspace the
preedit is still longer than the limit, it should still be
styled. This worked only if after typing backspace there were still
candidates. When there were no candidates, the preedit styling was not
applied even though the preedit was still long enough.  This is fixed
by adding the missing _update_preedit().