Add a property menu for the highest priority dictonary
This is mostly useful when using speech recognition input as the
language of the dictionary with the highest priority is used, if
possible, for speech recognition.

Therefore, one might want to switch the priority of dictionaries
frequently if one uses speech recognition for several languages.  I
already added key bindings to change the priorities of dictionaries.
Adding a property menu for the highest priority dictionary is
especially useful because one can then see which dictionary currently
has the highest priority without opening the setup tool.

Especially on non-Gnome desktops, when the floating panel is used, it
makes it easy to see which dictionary currently has the highest
priority.

When not using speech recognition, the priority of the dictionaries
has some influence, but it is not nearly as important as for speech
recognition, therefore I did not add such a menu until now.

Adding such a menu forced me to limit the number of possible
dictionaries to use a the same time to something because of
limitations in the property menus. Before, the number of dictionaries
was unlimited. I did choose 10 as the maximum number of dictionaries,
same as the maximum number of input-methods. That should not be a big
problem, 10 dictionaries at the same time is really a lot and one will
certainly already experience some slowdown.