Pressing Shift should not immediately change the case mode of the candidate
Only when Shift is released again and no other key was pressed inbetween
the press and the release of Shift should the case mode be changed.

The previous behaviour caused a problem with the default key binding
for 'select_previous_candidate' which is Shift+Tab: The initial press
of Shift already did the case mode change and then the following Tab
caused the selection of the previous candidate. But than that
candidate already had the case mode changed. Shift+Tab should just
select the previous candidate and not change the candidates.