Add option to automatically select the best candidate
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/44

What you type will automatically be corrected to the best
candidate by selecting the best candidate automatically. If
this option is off you have to select a candidate manually by
using the key binding to select the next candidate. If this
option is on and you do not want to use the automatically
selected candidate, you can use the key binding for the
“cancel” command which will deselect all candidates.