If a candidate is shown inline and *manually* selected, show the caret at the end of the candidate
If a candidate is shown inline:

• candidate not selected or automatically selected:
      Show caret at the  end of the typed string
• candidate manually selected:
      Show caret at the end of the candidate

Until now the caret was always shown at the end of the typed string which
is quite confusing if the candidate is already selected.

If the candidate is already *manually* selected, I made moving around
in the preëdit impossible anyway. Therefore, showing the caret
somewhere in the middle of the preëdit is useless. But showing the
caret at the end of the candidate is helpful, especially when preëdit
styling does not work like on Wayland.  It makes it more visually
obvious the candidate is selected, even when preëdit styling does not
work.

If a candidate is shown inline and selected, show the caret at the end of the candidate