Right, Left, BackSpace, and Delete edit the preëdit only if no candidate is *manually* selected
If a candidate is already *manually* selected, it is more natural if these keys
trigger a commit.

If no candidate is selected or the candidate has been *automatically*
selected because the option “Automatically select the best candidate”
is on, these keys continue to edit the preëdit.

Right, Left, BackSpace, and Delete edit the preëdit only if no candidate is *manually* selected