Skip some unittests when dictionaries or python modules needed are not installed
Some tests require dictionaries which are not available on all systems.
For example ko_KR.dic seems not to be available on Arch Linux and FreeBSD.
mr_IN.dic seems not to be available on Ubuntu and Elementary OS.

The Korean test was already skipped when the dictionary was not available
but without printing a nice message into the log file.

Several tests require the python3-enchant module to be installed, skip
these when the enchant module is unavailable.