When checking whether to show inline completion, normalize the first candidate as well
Normalizing only the typed string but not the first candidate was a bug
because this made the check whether the first candidate is a direct completion
of the typed string fail sometimes when accented characters were involved.

For example, if the typed string was

    Glühw

in NFC and the first candidate was

    Glühwürmchen

in NFD, then the first candidate was not recognized as a direct completion
of the typed string.