Extend inline completions automatically
If the first candidate is exactly the same as the typed string
prefer longer candidates which start exactly with the typed
string. If the user wants the typed string, he can easily
commit the preëdit, there is no need to select a candidate in
that case. Offering longer completions instead may give
the opportunity to save some key strokes.