Add support for (almost) arbitrary dead key sequences
Suggested by Marko Myllynen.

Marko thought it would be nice to be able to type

    ė̄  U+0117 LATIN SMALL LETTER E WITH DOT ABOVE U+0304 COMBINING MACRON

easily. This chracter is used in Samogitian:

    https://en.wikipedia.org/wiki/Samogitian_dialect#Writing_system

Marko suggested that writing it with a dead key sequence like

   <dead_abovedot> <dead_macron> <e>

might be a nice way of doing this as he already has these two dead keys
on his Finnish keyboard layout.

But this dead key sequence is not defined in any of the Compose files
in Xorg. And, Xorg doesn’t support results of compose sequences to be
longer than a single Unicode code point. And here we have already two
code points. Ibus and ibus-typing-booster both support compose
sequences with longer results, but adding such sequences to the Xorg
Compose files when they are not supported there seems wrong.
On top of that, there are many other such accented Latin characters
which cannot be represented by a single Unicode code point, should
we really add *all* of them to the Compose files? So Marko’s suggestion
was that it would be nice to handle all such dead key sequences
“reasonably”, even if they are not defined in the Compose files at all.

This commit tries to do this. If a sequence is defined in a Compose
file, it is always interpreted as defined in the file. But if
ibus-typing-booster encounters a dead key sequences which is not
defined in a Compose file but otherwise appears to be “reasonable”, it
tries to produce a “reasonable” result for that sequence which is
probably better than just throwing it away as an invalid sequence only
because it is not defined in any Compose file.