Skip emoji tests if emoji annotations are loaded from outside of the ibus-typing-booster source tree
On Fedora, there is a package cldr-emoji-annotation which contains the
emoji annotation xml files. Therefore, the ibus-typing-booster package
omits these files and requires the cldr-emoji-annotation package
instead.  But the files in the cldr-emoji-annotation package are
usually somewhat older than the files in the ibus-typing-booster
source tree, therefore some tests which are already adapted to the
newest versions of these files may fail when older versions are
used. Therefore, it is better to skip these tests when annotation
files outside of the ibus-typing-booster source tree are used.