Use update_gsettings=False when changing settings in test_itb.py
It is quite annoying when I run the test suite on my workstation and
the test suite changes the settings of the typing-booster I am using
for something else while the test suite runs. In test_gtk.py, this is
unavoidable, but in test_itb.py changing the gsettings is not
necessary.