Remove redundant arguments from super() calls in TypingBoosterEngine class
These arguments are not only redundant, they cause tracebacks during
unit testing since I changed to the better system of patching the IBus
classes with the mock classes. Because with the newer system I can add
the patches and also remove them again.

When gsettings change while a unit test is running, the callback
on_gsettings_value_changed() is called which may trigger some
of the super() calls. If the super() calls have no arguments, they
work no matter whether the mock patches are applied or not.
But with the redundant arguments, the super() calls may fail
depending on whether the patches are applied or not.