Save default dictionaries and input methods when making the lists empty in the setup tool
When one makes the list of dictionaries or input methods empty in the
setup tool, these lists automatically reset to the defaults for the
current locale.  But that default was not saved to the
settings. Therefore, the change was not immediately effective in
ibus-typing-booster and became only effective after “ibus restart”.