Remove the special candidate which is shown when a dictionary is missing
The setup tool now already shows nicely when a dictionary is missing.
This special candidate warning that a dictionary is missing is a
relict from a time when dictionaries were hardcoded and the setup tool
didn’t obviously show when a dictionary was missing. So this extra
warning is not needed anymore.  And it is just annoying when one
selects a language where the spellchecking dictionary is missing on
purpose, for example if one selects “en” instead of “en_US” or “en_GB”
on purpose because one wants to use the emoji and Unicode symbols but
not the English spellchecking dictionary.

See: https://github.com/mike-fabian/ibus-typing-booster/issues/98