Show the list of dictionaries in the setup tool only if it is more than just ['None']
['None'] is a special case because it indicates that the user really
wants to use no dictionaries at all. For the user it is most natural
to see the list of dictionaries as empty then.

Internally there is a difference between ['None'] and [''], an empty list means
that nothing has been set and defaults have been applied and ['None'] means
that the user has removed all dictionaries in the setup tool on purpose.

If one sets dictionary lists containing 'None' from the command line, for example:

$ dconf write /org/freedesktop/ibus/engine/typing-booster/dictionary '"None,en_GB,None,en_GB,de_DE"'

All these entries are shown in the list in the setup tool, even the
'None' entries.  So one can try such weird settings using the command
line (This is only useful for testing that even with such weird
settings nothing crazy happens, no crash, ...)

When using the setup tool, it is impossible to setup duplicate dictionaries
and lists containing 'None' which are not just ['None'].