Improve matching of keybindings, consider IBus.ModifierType.MODIFIER_MASK
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/94

The problem occurs only under Gnome Xorg, not under Gnome Wayland.

I could reproduce the problem in Fedora 32 and Ubuntu 20.04 using Gnome Xorg.

It happens when one has several keyboard layouts.

When using the first keyboard layout selectable in the Gnome panel and
types a key without any modifiers, xev shows the state 0x0000.
When using the second keyboard layout, xev shows the state 0x2000.
When using the third keyboard layout, xev shows the state 0x4000.
When using the fourth keyboard layout, xev shows the state 0x0000 again.
So this repeats in groups of 3 keyboard layouts.

To fix this I improved the matching of key bindings to consider only
bits in “state” which are really modifier bits. And, as before, ignore
the modifiers Caps Lock and Num Lock.

There maybe bits in the state mask which have nothing to do at all
with modifiers, they should all be ignored when matching keys.