Fix wrong indentation of push_context() in _commit_string()
This error was only causing a problem for clients where surrounding
text is not supported at all, for example in a client like xterm. In
such cases, the context was never pushed. At the start the context is
of course always empty. Therefore, the push_context() in the “if
self.get_phrase() ..”  block was never executed in xterm.