The keybindings in _fill_keybindings_edit_popover_listbox() should not be sorted
Currently this doesn’t matter in ibus-typing-booster.

But in ibus-table, keybindings to commit candidates from the lookup table
are used to label the candidates in the lookup table. If there is more
than one keybinding to commit a candidate from the lookup table, only
the first one is used as a label. I.e. the order of these keybindings
may matter, better don’t sort them.