Add a command 'next_case_mode' with configurable key bindings
Changing the case mode was hard coded to Shift_L and Shift_R.

Modifier keys like Shift_L and Shift_R have to be matched using both
the press and the release event and making sure there was nothing in
between.

This didn’t work using the “class HotKeys” from itb_util.py, therefore
the matching of Shift_L and Shift_R was hardcoded as a special case
and as it was hardcoded it was not possible to configure these key
bindings.

But recently “class HotKeys” was improved to handle matching of such
modifier keys as well. So now a command 'next_case_mode' has been
added with configurable keybindings (default ['Shift_L', 'Shift_R']).