Prevent also Delete from reopening a preedit when the option “Arrow keys can reopen a preedit” is off
Reopening the preedit is so terribly buggy, especially under Gnome Wayland
but also in many other applications.

Often surrounding text is reported as supported by the application but
ibus-typing-booster then gets completely wrong results when querying
the surrounding text. Without properly working surrounding text,
reopening a preedit cannot work correctly either.

Until now this option switched reopening the preedit on and off only for the arrow keys.
So even when this was switched off for the arrow keys, it was still causing problems
when using backspace in case of broken support for surrounding text.

Now reopening preedits is disabled completely when this option is off, which is
better in most cases as surrounding text is so terribly broken.

One can still turn it on, it might be helpful in some cases where
surrounding text works reasonably well (e.g. in gedit when using Xorg
and not Wayland). But in most cases it is unfortunately better to
switch this off.