If a commit string contains new-lines, commit it in several chunks and forward Return between the chunks
See: https://github.com/mike-fabian/ibus-typing-booster/issues/158#issuecomment-794233274

When committing a string which contains newlines while writing a new
mail using gmail in Firefox, the newlines appear as spaces in the
GMail editor. Apparently the same happens in Thunderbird. In other
cases, like when writing a new mail with GMail in google-chrome, it
works.

Using several commits to commit the multi-line string line by line doesn't change this behaviour at all, the newlines are still converted to spaces when writing a new mail in GMail in firefox.

super().commit_text(IBus.Text.new_from_string('\n'))

gives a space, not a new line when writing a new mail in GMail in firefox.

But, forwarding a Return key event instead of committing a '\n' apparently works
always.

Therefore, as a workaround, this change commits a multi-line string
line-by line and forwards Return keys between the lines. A short sleep
is needed after each line otherwise all commits appear first and all
forwarded Returns appear after that. Ugly, but it seems to improve the
situation.