Don’t let dialog action area of setup tool expand.
Letting this expand just wastes space when one resizes the window of
the setup tool to make it bigger. Extra space going to the dialog
action area is rather useless it should better be used to make
scrolled windows like the multi line entry for the shortcuts or the
treeview of already defined shortcuts bigger.