Starting a compose sequence while a candidate is selected should append to the selection
When a regular letter key is typed while a candidate is selected, the selection
is put into the preedit and the typed letter is appended.

This should work the same way with compose sequences:
When a compose sequence is started while a candidate is selected, the selection
should be put into the preedit and the compose sequence appended.

Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/175