Completly new implementation of reopening the preedit on arrow-keys, BackSpace and Delete
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/196

This is still not perfect, but *much* better than the old implementation.
I could still make it fail when testing in rare cases, but it seems much
more reliable.

It became much more reliable because I now try to reopen the preedit
after the release event of the key happened, not after the press event.

When typing “hello ” followed by an Left (arrow-left), matching on the
press event means that the cursor is still behind the space so I need
to delete the “hello” with delete_surrounding_text() *and* move the
cursor left over the space.  So I need to do two actions which have an
effect on the surrounding text and which are sensitive to timing and
might even happen in the wrong order.

Now, I match the release event. When Left is released, the cursor
should already by behind “hello”, not behind the space. So in that
case it is enough to delete the “hello” with delete_surrounding_text()
and no cursor movement is nessesary. That already makes it much more
reliable.

Same for Right, Backspace, and Delete, I need only one
delete_surrounding_text() now for all of these.

In addition to that, I am trying now to detect whether the surrounding
text was properly updated. At the end of each do_process_key_event()
I clear an Event object and remember the old surrounding text:

    self._set_surrounding_text_event.clear()
    self._surrounding_text_old = self.get_surrounding_text()

When a 'set-surrounding-text' signal occurs, the function
on_set_surrounding_text() connected to that signal is called
and sets this Event object:

    self._set_surrounding_text_event.set()

After a Left (arrow-left) is pressed, this Event object is cleared at
the end of do_process_key_event().  Then, wenn Left is released again,
do_process_key_event() is called again and on key release events this
calls _maybe_reopen_preedit(key).

This function can check wether a 'set-surrounding-text' signal
happened since the Left press event:

   self._set_surrounding_text_event.wait(timeout=0.1)

If such an event has happened already, this continues immediately.  If
not, this waits up to 0.1 seconds for such an event to happen.
If the event still has not happened after 0.1 seconds,
this check detects that:

        if not self._set_surrounding_text_event.is_set():
            LOGGER.debug(
                'Surrounding text has not been set since last key event. '
                'Something is wrong with the timing. Do not try to reopen '
                'the preedit.')
            return False

and in that case I can skip reopening the preedit and return False
because something is clearly out of whack, I cannot trust the results
I get from get_surrounding_text() and I might end up putting something
completely wrong back into preedit and delete_surrounding_text() might
delete something completely wrong as well.

Then I compare how the new surrounding text when the key release event
is processed differs from the old surrounding text at the end of the
processing of the press event and do some sanity checks:

- if it was a Left or KP_Left or BackSpace press and release, the cursor
  in the surrounding text should have moved *exactly* one column left.
- If it was a Right or KP_Right press and release, the cursor
  in the surrounding text should have moved one *or* *more* columns
  right.
- If it was a Delete or KP_Delete press and release, the cursor
  in the surrounding text should not have moved *at* *all*.

If any of these sanity checks fails, I return False and do *not*
attempt to reopen a preedit.

This new implementation sometimes fails to reopen a preedit when
it appears that this should have been possible.

In these cases, something went wrong with surrounding text, it was not
updated or the update was wrong. If this happens, it is much better to
return False and do nothing instead trying to reopen a preedit with
wrong data and create a mess.