Never mix context from surrounding text with remembered context
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/204

Mixing context determined from surrounding text with context
remembered seems to do more harm than good. Better no context or
shorter context than wrong context!

Unfortunately surrounding text seems to fetch only the current line.
That means if the cursor is near the beginning of the second line
of this example (| indicates the cursor position)

    This text uses
    two |lines

surrounding text fetches only “two lines” and the cursor position 4.
In that surrounding text, only the token “two” can be found to the
left of the cursor position.

After typing “This text uses” + Return + “two”, the partly remembered
context “text uses” plus the new context “two” resulted in the total
context “text uses two” which happens to be correct.

But it easily becomes incorrect if the cursor is moved around in the
second line. When moving the cursor back to the beginning of the
second line, i.e.

    This text uses
    |two lines

The remembered context is “text uses two” and the context from
surrounding text is empty. But that context is incorrect for the
new cursor positon, the “two” should not be in the context.

If no surrounding text can be used at all, remembering the context is
a fallback which is “better than nothing”.

But if surrounding text is used, mixing tokens determined from
surrounding text with remembered tokens seems to cause more harm than
good. It might still be correct if one types several lines without
ever repositioning the cursor and it that special case it might be
helpful, but as soon as cursor repositioning near the beginning of a
line is involved, mixing remembered with new context is likely to be
completely wrong.

Better don’t mix this and hope that surrounding text can be improved to look
back to the end of the previous line in future.