When setting self._input_mode to False (direct input), set self._hide_input to False as well
If self._hide_input was True while self._input_mode was True and
then self._input_mode was set to False, self._hide_input mode stayed True.
This had no effect on normal input but it did hide the compose preview with ***.

As the hotkey for 'toggle_hide_input' is not matched when
self._input_mode == False there is no way to switch self._hide_input
to False again while self._input_mode == False.  Theoretically I could
match the 'toggle_hide_input' as well while self._input_mode == False
but that might have side effects because the keybinding for that
command would then not be available for other things while in direct
input mode. The default keybinding for 'toggle_hide_input' is
['Shift+Control+Escape'] and that key would not be available more,
even in direct input mode.  If some application uses exact that
keybinding for something, it might cause problems if
ibus-typing-booster takes that keybinding away.

Direct input mode should be “as direct as possible” and try not to interfere with
anything.

Therefore the only things done in direct input mode are the handling
of compose sequences and matching the keybinding for
'toggle_input_mode_on_off'.

Therefore, self.hide_input should either be ignored while self._input_mode == False
or set to False as well.

Setting self.hide_input to False as well of course means when
switching ibus-typing-booster back on again it stays False and needs
to be switched to True again if desired. I think this is the more
useful behaviour as hiding input is intended to temporarily hide
something. Remembering this and keeping it on seems more surprising
then useful to me.