Improve behaviour of key combinations which normally select text when a preëdit is open
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/210

Usually, pressing Shift and any of the following keys:

    IBus.KEY_Right, IBus.KEY_KP_Right,
    IBus.KEY_Left, IBus.KEY_KP_Left,
    IBus.KEY_End, IBus.KEY_KP_End,
    IBus.KEY_Home, IBus.KEY_KP_Home,
    IBus.KEY_Down, IBus.KEY_KP_Down,
    IBus.KEY_Up, IBus.KEY_KP_Up,
    IBus.KEY_Page_Down, IBus.KEY_KP_Page_Down, IBus.KEY_KP_Next,
    IBus.KEY_Page_Up, IBus.KEY_KP_Page_Up, IBus.KEY_KP_Prior

selects text.

At least in Gtk text entries, Qt5 text entries, gedit, firefox, they do this.
(It is different in terminals like gnome-terminal and xterm, and probably
different in all programs using some old X11 toolkit ...)

Until now, behaviour of these "text selecting" key combinations while
a preedit was open was not intuitive and differed between key
combinations.  Some like Shift+Left kept editing the preedit just like
only Left.  Others like Shift+Home did commit the preedit and continue
to select something, but if the cursor happened to be not at the end
of the preedit when pressing the key combination, then the selected
text area was wrong because it always used the right end of the
preedit as the cursor position to select from even if the cursor was
somewhere in the middle of the preedit.

Now all the above key combinations with Shift do this:

- commit the preedit immediately
- forward Left key events to correct the cursor position if needed
- pass the key combination which triggered the commit to the application
  which now can select from the right cursor position.