Rename option “Arrow keys can reopen a preedit” to “Enable reopening preedits”
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/212

The name of the option does not really fit anymore because it is not only
affecting the arrow keys.

Currently, these keys can cause a preedit to be reopened:

    IBus.KEY_Left, IBus.KEY_KP_Left, IBus.KEY_Right, IBus.KEY_KP_Right,
    IBus.KEY_BackSpace, IBus.KEY_Delete, IBus.KEY_KP_Delete

I may add other keys to that list and maybe even allow reopening of
preedits after movements of the cursor with the mouse.

The gsettings key is still unchanged (“arrowkeysreopenpreedit”).

I will change that later and maybe change the default to True.