get_string_from_lookup_table_cursor_pos() should return the string in NFC
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/234

The problem occured because after typing “discrep”,
then selecting the candidate “discrepância” in the lookup table and then
typing “,”, the candidate returned by get_string_from_lookup_table_cursor_pos()
was in NFD, i.e. “discrepância” was put into self._typed_string in NFD.
After typing a “,” there was “discrepância,” in NFD in self._typed_string.
So len(self._typed_string) was 14 because of NFD.
get_caret() then calculated the caret position as 13 using NFC.

When a space was typed in this situation, the string to commit was
taken from input_phrase which came from self._typed_string and was
NFD. So commit_string in the following code had a length of 14 but
caret_was calculated from the length of input_phrase in NFC was 13. So
one nonsensical Left event was forwarded:

            caret_was = self.get_caret()
            self._commit_string(commit_string, input_phrase=input_phrase)
            if not candidate_was_selected:
                # cursor needs to be corrected leftwards:
                for dummy_char in commit_string[caret_was:]:
                    self._forward_generated_key_event(IBus.KEY_Left)

which did put the cursor before the “,” and then the space was forwarded.
So the final result was “discrepância ,”.

Making get_string_from_lookup_table_cursor_pos() return NFC removes this
inconsistency.

When a candidate selected in the lookup table is moved into the
self._typed_string because one continues typing, using NFC is also
better for other reasons.  The preedit is always displayed in NFC
anyway. When moving the cursor around in the preedit and
self._typed_string contains NFD, one has to type two times Left to
move the cursor backwards over an “â”. If the “â” has been typed using
something like t-latn-post, this is unavoidable because then
self.typed_string contains ['a', '^'] which should not be normalized.

But when self.typed_string is updated with a candidate from the lookup
table, that candidate should be normalized to NFC.