Fix “Inconsistency when typing digits in Compose #238”
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/238

Hotkey handling matches while a compose sequence is in progress does this:

If the key is a valid key to continue the compose sequence,
abandon hotkey matching and return false.

Else remove the hotkey from the compose sequence and try
to execute the command bound to the hotkey.

Now if the command bound to the hotkey succeeds (returns True), it is
OK that the hotkey was removed from the compose sequence and just
returning True is fine.

For example the hotkey might have been Down, it was handled by
selecting the next candidate in the compose completion lookup table,
it should not be in the compose sequence, so it was OK to remove
that key.

But if the command bound to the hotkey does not succeed (returns
False), the key was actually not handled and so the key needs to be
put back into the compose sequence.