Add 'G-_' to 'G- ' as exceptions which must not trigger commits but insert into the preedit
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/243
Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=2002199

The German neo2 layout uses ISO_Level3_Shift+v (= CapsLock+w on a QWERTY layout)
to type an underscore '_'. That means when an underscore is typed using
that layout, key.msymbol is 'G-_'. This triggered a commit instead of inserting
the '_' into the preedit which made it impossible to add a '_' to the preedit
to trigger the emoji mode.

This commit fixes this problem by making another exception for 'G-_'.