Skip emoji tests which rely on the enchant results if enchants results changed
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/244

It is good enough if these tests run on my development system.
If the behaviour of enchant changes on my development system, I’ll adapt
the test cases.

But when Typing Booster is build on other distributions with different versions
of enchant and/or different versions of dictionaries, these tests may fail
for no good reason. So better skip these tests when the enchant output seems
different than on my system.