Make matching in the database case insensitive
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/251

Resolves the second part of that bug, a previous commit resolved it for
matching in the dictionaries already.

This commit here uses the same trick as for accent insensitive
matching in the database: Store only lower case in “input_phrase” and
convert everything to lower case before matching against
“input_phrase”.

“PRAGMA case_sensitive_like = false;” did cause a slowdown of a factor of 2
in my tests even when the database had only 50000 rows. The advantage of using
this would be that one could add an option to switch between case insensitive
and case sensitive matching which could have an effect immediately.

Doing the case insensitive match by storing only lower case in the database
and use .lower() before trying to match agains “input_phrase” does not seem
to cause any measurable slowdown at all. The disadvantage is that it has an
effect only on new rows in the database. I.e. if an option were added to switch
between case insensitive and case sensitive matching, that option would
only have an effect of rows added to the database after switching that option.

But maybe I don’t add such an option at all and even if I add such an
option, nobody would switch that option all the time, i.e. the problem
that it effects only new entries would fade away in the long run
because of expiry of old database entries. But the speed advantage is
there forever. Therefore, I think not using the above PRAGMA is better.