Rewrite some doctests depending on spellcheckers
Results of these doctests depend on whether pyhunspell or
python3-enchant is used.  And in case of python3-enchant the provider
may be aspell, hunspell, myspell, ispell ...

So depending on what exactly is installed on the system, spell checking results
may be different.

Therefore, I rewrote some doctests to be more generic and hopefully work
with all spellcheckers.