Do not record stuff to user database when itb_util.InputHints.PRIVATE is set
Currently this does nothing because GTK_INPUT_HINT_PRIVATE is
currently only in Gtk4 and not in Gtk3:

https://docs.gtk.org/gtk3/flags.InputHints.html
https://docs.gtk.org/gtk4/flags.InputHints.html

And ibus-typing-booster uses Gtk3 at the moment, so itb_util.InputHints.PRIVATE
will never be set.

But as soon as this is added to ibus and Gtk3 this should start working.

Or in future if everything is ported to Gtk4 it should start working.