Allow capital letters and accents in user shortcuts
If normal words are typed which are not shortcuts, they are always
matched case and accent insensitively in the database. I did some
statistics and the results showed that doing this without considering
case and capitalization improves the predictions on average (i.e. it
improves on average the number of letters which can be saved typing).

But for shortcuts this is a different thing, in shortcuts, I could
allow accents and capital letters without making anything worse,
giving the user a few moreoptions to define distinct shortcuts.

At the moment I can see no disadvantages in allowing this in
shortcuts. Except that if I allow this now and reconsider later and
disallow them later, users which in the meantime already defined
shortcuts with capital letters and/or accents would loose these
shortcuts again.