Add a “Restore all defaults” button to the setup tool
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/272

I made an exception for two keys:

The 'googleapplicationcredentials' stores a filename for the credentials.
Resetting this to the default, which is the empty string ('') cannot fix
or improve anything.

Similarly, setting 'dictionaryinstalltimestamp' to the empty string
seems useless. 'dictionaryinstalltimestamp' is not settable directly
by the user anyway, it is only set when new dictionaries were
installed using the “Install missing dictionaries” button in the setup
tool

Therefore, I skip these two keys in the loop resetting the keys.