Make it possible to use inline completion without automatic fallback to a candidate list
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/266

Now the inline completion option has 3 values:

• No
• Yes, with fallback to popup
• Yes, without fallback to popup

When choosing “Yes, with fallback”, the behaviour is as it used to be
when the old checkbox was on: the completion is shown inline if this
is possible without changing the already typed string in the preedit,
if this is not possible, a popup candidate list is automatically shown
instead as a fallback.

When choosing “Yes without fallback to popup”, a popup candidate list
is never shown automatically as a fallback but only when explicitely
requested by typing the key bound to the command
“select_next_candidate” (By default this is the Tab key).

When updating to the new release with this new option, the value of
the old checkbox gets lost. I.e. no matter whether the user had the
inline completion option checked or not checked before the update,
after the update the value for “inline completion” reverts to the
default “No”. So if one was using “inline completion” before the
update, one has to switch it on again after the update.