Add a command 'commit' with configurable key bindings
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/320

Useful to commit the current input without adding any keys (apart from
adjusting the cursor position, some Left events may be forwarded to
ajust the cursor position if the cursor was not at the right end of
the preedit when the commit happened).

For example, the Japanese input methods ibus-anthy, ibus-kkc,
ibus-mozc, ... commit when Return or Enter is typed and do not forward
the Return or Enter to the application.

Typing Booster also commits when Return or Enter is typed *and* then
forwards the Return or Enter to the application. This saves one
keystroke when typing Return or Enter at the end of the line. Because
it saves a keystroke it will remain the default, but now it is
possible to bind the command to these keys

    commit     ['Return', 'Enter']

and then the bound keys behaves like in the Japanese input
methods, cause a commit but don’t forward the key. If one uses this
and wants a new line while the preedit is active, one needs to type
Return twice, first to commit, second for the new line.

But this may be useful if one often wants to commit without adding
anything to the commit, for example if one wants to type several emoji
or symbols in a row without adding spaces in between the emoji, then a
key to commit without adding anything might be useful. Some users use
Typing Booster mostly or even only for emoji and symbols, in that case
having a dedicated key to commit without forwarding the key might be
useful, even if it is the Return or Enter key and one has to type it
twice at the end of the line.

Personally I often use 'Right' (rightwards arrow) to commit without
adding anything, but this only works at the end of the current buffer.
When committing with 'Right', a 'Right' key event is send after the
commit but if one was typing at the end of the buffer, this 'Right'
event does nothing.

Of course one could bind only the Return or only the Enter key to the
'commit' command

    commit     ['Return']
    commit     ['Enter']

Then one still has the other one to commit *and* add a new line.

Or one could bind a completely different key to 'commit' ...